Additions to the migration notes, copyrights, deprecations for v3.0
