+++ Research/Research/PlatformContext/RSDResourceInfo+Platform.swift
@@ -1,8 +1,8 @@
 //
-//  RSDUIAction+Platform.swift
-//  ResearchPlatformContext
+//  RSDResourceInfo+Platform.swift
+//  Research
 //
-//  Copyright © 2019 Sage Bionetworks. All rights reserved.
+//  Copyright © 2017-2019 Sage Bionetworks. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without modification,
 // are permitted provided that the following conditions are met:
@@ -31,44 +31,19 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
 
-#if os(macOS)
-import AppKit
-#else
-import UIKit
-#endif
+import Foundation
 
-extension RSDUIAction {
+extension Bundle : RSDResourceBundle {
+}
+
+extension RSDResourceInfo {
     
-    /// Convenience property for accessing the button icon.
-    @available(*, deprecated, message: "Use `RSDImageRules.assetImage()` instead.")
-    public var buttonIcon: RSDImage? {
-        guard let imageName = self.iconName else { return nil }
-        #if os(iOS) || os(tvOS)
-        return RSDDesignSystem.shared.imageRules.assetImage(named: imageName,
-                                                            using: self,
-                                                            compatibleWith: nil)
-        #elseif os(macOS)
-        return RSDImage(named: .init(imageName))
-        #else
-        return RSDImage(named: imageName)
-        #endif
+    /// The bundle returned for the given `bundleIdentifier` or `factoryBundle` if `nil`.
+    public var bundle: Bundle? {
+        guard let identifier = bundleIdentifier
+            else {
+                return self.factoryBundle as? Bundle
+        }
+        return Bundle(identifier: identifier)
     }
-    
-    #if os(iOS) || os(tvOS)
-    
-    /// Returns the button icon (if found).
-    ///
-    /// - parameters:
-    ///     - resourceName: The name of the image.
-    ///     - resourceInfo: The resource info associated with this asset.
-    ///     - traitCollection: The trait collection for the presenting view.
-    /// - returns: Image if found.
-    public func buttonImage(using designSystem: RSDDesignSystem? = nil,
-                    compatibleWith traitCollection: UITraitCollection? = nil) -> RSDImage? {
-        guard let imageName = self.iconName else { return nil }
-        let imageRules = designSystem?.imageRules ?? RSDDesignSystem.shared.imageRules
-        return imageRules.assetImage(named: imageName, using: self, compatibleWith: traitCollection)
-    }
-    
-    #endif
 }
