+++ Research/Research/RSDTaskMetadata.swift
@@ -2,7 +2,7 @@
 //  RSDTaskMetadata.swift
 //  Research
 //
-//  Copyright © 2018 Sage Bionetworks. All rights reserved.
+//  Copyright © 2018-2019 Sage Bionetworks. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without modification,
 // are permitted provided that the following conditions are met:
